created new error enums
