+++ QuEST/CPU/QuEST_cpu.c
@@ -13,6 +13,9 @@
 
 # include "QuEST_cpu_internal.h"
 
+// debug: remove this after all validation is removed
+# include "../QuEST_validation.h"
+
 # include <math.h>  
 # include <stdio.h>
 # include <stdlib.h>
